+++ unit_tests/Stoppable/TaskTests.cc
@@ -1,0 +1,94 @@
+#include "ExceptionPointee.hpp"
+#include "Task.hpp"
+
+#include <string>
+#include <thread>
+
+namespace Stoppable::tests {
+using namespace std;
+using namespace ::testing;
+
+class TaskTests : public Test {
+protected:
+  void testCanStart() {
+    EXPECT_FALSE(task_->running());
+
+    EXPECT_NO_THROW(task_->start());
+    EXPECT_TRUE(task_->running());
+  }
+
+  void testCanStartAndStop() {
+    EXPECT_NO_FATAL_FAILURE(testCanStart());
+
+    EXPECT_NO_THROW(task_->stop());
+    EXPECT_FALSE(task_->running());
+  }
+
+  MockFunction<void()> mock_cycle_;
+  MockFunction<void(const std::exception_ptr&)> mock_handler_;
+  TaskPtr task_ = make_shared<Task>(
+      mock_cycle_.AsStdFunction(), mock_handler_.AsStdFunction());
+};
+
+TEST_F(TaskTests, canStartAndStop) {
+  EXPECT_CALL(mock_handler_, Call(_)).Times(Exactly(0));
+  EXPECT_CALL(mock_cycle_, Call()).Times(AtLeast(1));
+
+  EXPECT_NO_FATAL_FAILURE(testCanStartAndStop());
+}
+
+TEST_F(TaskTests, canStartAndStopOnDestruction) {
+  EXPECT_CALL(mock_handler_, Call(_)).Times(Exactly(0));
+  EXPECT_CALL(mock_cycle_, Call()).Times(AtLeast(1));
+
+  EXPECT_NO_FATAL_FAILURE(testCanStart());
+  task_.reset();
+}
+
+TEST_F(TaskTests, canStartTwice) {
+  EXPECT_CALL(mock_handler_, Call(_)).Times(Exactly(0));
+  EXPECT_CALL(mock_cycle_, Call()).Times(AtLeast(1));
+
+  EXPECT_NO_FATAL_FAILURE(testCanStart());
+  EXPECT_NO_THROW(task_->start());
+  EXPECT_TRUE(task_->running());
+  EXPECT_NO_THROW(task_->stop());
+  EXPECT_FALSE(task_->running());
+}
+
+TEST_F(TaskTests, canStopTwice) {
+  EXPECT_CALL(mock_handler_, Call(_)).Times(Exactly(0));
+  EXPECT_CALL(mock_cycle_, Call()).Times(AtLeast(1));
+
+  EXPECT_NO_FATAL_FAILURE(testCanStartAndStop());
+  EXPECT_NO_THROW(task_->stop());
+  EXPECT_FALSE(task_->running());
+}
+
+TEST_F(TaskTests, canRestart) {
+  EXPECT_CALL(mock_handler_, Call(_)).Times(Exactly(0));
+  EXPECT_CALL(mock_cycle_, Call()).Times(AtLeast(2));
+
+  EXPECT_NO_FATAL_FAILURE(testCanStartAndStop());
+  // restart task
+  EXPECT_NO_FATAL_FAILURE(testCanStartAndStop());
+}
+
+TEST_F(TaskTests, canHandleException) {
+  auto target_exception = logic_error("Test thrown exceptions");
+
+  EXPECT_CALL(mock_handler_, Call(ExceptionPointee(target_exception)))
+      .Times(Exactly(1));
+  EXPECT_CALL(mock_cycle_, Call())
+      .Times(AtLeast(1))
+      .WillOnce(Throw(target_exception))
+      .WillRepeatedly([]() {
+        // default action for valgrind, not normally used in the test
+        // @see RoutineTests.canThrowAndRun test case for explanation
+        this_thread::sleep_for(100ms);
+      });
+
+  EXPECT_NO_FATAL_FAILURE(testCanStartAndStop());
+}
+
+} // namespace Stoppable::tests